detect case problems, fix 1, analyzer for embed
