+++ package.json
@@ -6,6 +6,7 @@
     "@babel/preset-env": "^7.8.3",
     "autoprefixer": "^9.7.3",
     "babel-jest": "^24.9.0",
+    "case-sensitive-paths-webpack-plugin": "^2.4.0",
     "circular-dependency-plugin": "^5.2.0",
     "concurrently": "^5.3.0",
     "cross-env": "^5.2.0",
