+++ webpack.embed.config.js
@@ -1,10 +1,11 @@
 const baseConfig = require("./webpack.base.config");
 
+
 const targets = [
-  ["./src/embed/enhance.js", "/public/embed", "enhance.js"],
-  ["./src/embed/noteLoader.js", "/public/notes", "loader.js"],
-  ["./src/embed/documentLoader.js", "/public/viewer", "loader.js"],
-  ["./src/embed/projectLoader.js", "/public/embed", "loader.js"],
+  ["./src/embed/enhance.js", "/public/embed", "enhance.js"], /* page embeds */
+  ["./src/embed/noteLoader.js", "/public/notes", "loader.js"], /* node embed */
+  ["./src/embed/documentLoader.js", "/public/viewer", "loader.js"], /* document embed */
+  ["./src/embed/projectLoader.js", "/public/embed", "loader.js"], /* was search embed, but now only project embeds */
 ];
 
 module.exports = targets.map(([src, outputPath, outputFilename]) => ({
